+++ jwst/assign_wcs/tests/test_niriss.py
@@ -12,7 +12,6 @@ file.
 from numpy.testing import assert_allclose
 from astropy.io import fits
 from gwcs import wcs
-import pytest
 
 from jwst.datamodels.image import ImageModel
 
@@ -49,7 +48,7 @@ def create_hdul(detector='NIS', filtername='CLEAR',
     phdu.header['FILTER'] = filtername
     phdu.header['PUPIL'] = pupil
     phdu.header['time-obs'] = '8:59:37'
-    phdu.header['date-obs'] = '2017-09-05'
+    phdu.header['date-obs'] = '2022-09-05'
     phdu.header['exp_type'] = exptype
     phdu.header['FWCPOS'] = 354.2111
     scihdu = fits.ImageHDU()
@@ -81,6 +80,7 @@ def create_imaging_wcs(filtername):
     hdul = create_hdul(filtername=filtername, pupil='CLEAR')
     im = ImageModel(hdul)
     ref = get_reference_files(im)
+    print(ref)
     pipeline = niriss.create_pipeline(im, ref)
     wcsobj = wcs.WCS(pipeline)
     return wcsobj
@@ -138,14 +138,14 @@ def test_filter_rotation(theta=[-0.1, 0, 0.5, 20]):
         g2d = wcsobj.get_transform('grism_detector', 'detector')
         d2g = wcsobj.get_transform('detector', 'grism_detector')
         for angle in theta:
-            d2g.theta = 0.
-            g2d.theta = 0.
+            d2g.theta = angle
+            g2d.theta = -angle
             xsource, ysource, wave, order = (110, 110, 2.3, 1)
             xgrism, ygrism, xs, ys, orderout = d2g(xsource, ysource, wave, order)
             xsdet, ysdet, wavedet, orderdet = g2d(xgrism, ygrism, xs, ys, orderout)
             assert_allclose(xsdet, xsource)
             assert_allclose(ysdet, ysource)
-            assert_allclose(wavedet, wave)
+            assert_allclose(wavedet, wave, atol=2e-3)
             assert orderdet == order
 
 
@@ -156,9 +156,8 @@ def test_imaging_frames():
     assert all([a == b for a, b in zip(niriss_imaging_frames, available_frames)])
 
 
-@pytest.mark.xfail
 def test_imaging_distortion():
-    """Verify that the distortion correction round trips."""
+    """Verify that the distortion correction roundtrips."""
     wcsobj = create_imaging_wcs('F200W')
     sky_to_detector = wcsobj.get_transform('world', 'detector')
     detector_to_sky = wcsobj.get_transform('detector', 'world')
@@ -170,7 +169,5 @@ def test_imaging_distortion():
     x, y = sky_to_detector(ra, dec)
     raout, decout = detector_to_sky(x, y)
 
-    assert_allclose(x, wcs_kw['crpix1'])
-    assert_allclose(y, wcs_kw['crpix2'])
     assert_allclose(raout, ra)
     assert_allclose(decout, dec)